[Reader] KImage: Take care of long-strip images
Long vertical images are commonly used in Webtoons.
They must be subdivided so as not to exceed the hardware limit.

I don't currently know how to determine the limit imposed by the 
hardware and the software (GDK?).
The value used is therefore deliberately very low in order to suit the 
majority.